requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.0.3
python-dotenv==1.0.0
requests==2.31.0
selenium==4.24.0
webdriver-manager==4.0.2
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0
